+++ src/router/index.js
@@ -30,6 +30,10 @@ import AdminOSP from '../views/Admin/AdminOSP.vue'
 import AdminSWDA from '../views/Admin/AdminSWDA.vue'
 import AdminSWDACreate from '../views/Admin/CRUD_Pages/AdminSWDACreate'
 import AdminSWDAEdit from '../views/Admin/CRUD_Pages/AdminSWDAEdit'
+import AdminSWDAView from '../views/Admin/CRUD_Pages/AdminSwdaView'
+
+
+import SettingDashboard from '../views/Admin/Settings_Pages/SettingDashboard'
 
 import ERROR from '../views/404.vue' // Import the 404 Error Page
 import { userIsAuthenticated } from '../auth'; // Import the user authentication function
@@ -85,13 +89,13 @@ const routes = [
         // Allow access to the route
         next();
       }
-      else if(adminIsAuthenticated()){
-          // Admin is authenticated, clear the token in localStorage
-          sessionStorage.removeItem('admin');
-          // Allow access to the route
-          next();
+      else if (adminIsAuthenticated()) {
+        // Admin is authenticated, clear the token in localStorage
+        sessionStorage.removeItem('admin');
+        // Allow access to the route
+        next();
       }
-       else {
+      else {
         // User is not authenticated, proceed to the login page
         logout();
         next();
@@ -170,6 +174,15 @@ const routes = [
     beforeEnter: adminRequireAuth,
   },
 
+
+  {
+    path: '/adminswda/:ID/view',
+    name: 'AdminSWDAView',
+    component: AdminSWDAView,
+    beforeEnter: adminRequireAuth,
+  },
+
+
   {
     path: '/adminswda/:ID/edit',
     name: 'AdminSWDAEdit',
@@ -177,6 +190,13 @@ const routes = [
     beforeEnter: adminRequireAuth,
   },
 
+
+  {
+    path: '/adminSettings',
+    name: 'SettingDashboard',
+    component: SettingDashboard,
+    beforeEnter: adminRequireAuth,
+  },
 
   // ADMIN DASHBOARD END
 
@@ -263,7 +283,7 @@ const routes = [
 
   // USER DASHBOARD END
 
-    //catch-all route for 404 errors
+  //catch-all route for 404 errors
   {
     path: '/:catchAll(.*)', // Matches any URL
     name: 'error',