+++ src/views/User/SWDA.vue
@@ -178,7 +178,7 @@
         </div>
       </div>
 
-      <div class="agencies col-12 col-md-6 col-lg-5">
+      <div class="agencies col-12 col-md-12 col-lg-5">
         <div class="shadow">
           <h4>MODE OF DELIVERY</h4>
           <p class="col-3 textAgencies">Community</p>
@@ -215,7 +215,9 @@
           <br />
 
           <p class="col-12">
-            <span id="numberAgencies" class="numberAgencies">22</span>
+            <span id="numberAgencies" class="numberAgencies"
+              >{{ agencies.length }}
+            </span>
           </p>
         </div>
       </div>
@@ -307,8 +309,8 @@
             SECTORS
           </p>
           <DoughnutChart
-            v-if="ClusterData"
-            :data="ClusterData"
+            v-if="SectorData"
+            :data="SectorData"
             :legendOptions="customLegendOptions"
           />
         </div>
@@ -329,8 +331,8 @@
             CLIENT
           </p>
           <DoughnutChart
-            v-if="ClusterData"
-            :data="ClusterData"
+            v-if="ClientData"
+            :data="ClientData"
             :legendOptions="customLegendOptions"
           />
         </div>
@@ -380,6 +382,8 @@ export default {
       ClusterData: null, // Initialize Chart Data as null
       RegionData: null, // Initialize Chart Data as null
       ClienteleData: null,
+      SectorData: null,
+      ClientData: null,
 
       // Cluster Lengths
       southClusterLength: 0, // Number of agencies in the South Cluster
@@ -436,6 +440,7 @@ export default {
   },
 
   methods: {
+    // *ALREADY FUNCTIONAL
     ClusterFetchData() {
       return axios
         .get("http://127.0.0.1:8000/api/cluster")
@@ -549,6 +554,7 @@ export default {
           this.ClusterData = clusterdata;
         });
     },
+    // *ALREADY FUNCTIONAL
     RegionFetchData() {
       return axios
         .get("http://127.0.0.1:8000/api/regionaloperation")
@@ -650,6 +656,7 @@ export default {
           this.RegionData = regiondata;
         });
     },
+    // *ALREADY FUNCTIONAL
     AgencyFetchData() {
       return axios
         .get("http://127.0.0.1:8000/api/agencies")
@@ -723,33 +730,60 @@ export default {
           console.error("Error fetching data:", error);
         });
     },
-
-    //! NEED FIXINGS
-    ClienteleFetchData() {
+    // *ALREADY FUNCTIONAL
+    SectorFetchData() {
       return axios
-        .get("http://127.0.0.1:8000/api/clientele")
+        .get("http://127.0.0.1:8000/api/sector")
         .then((response) => {
-          // Initialize data arrays
+          // Initialize arrays for active and expired agencies
           const children = [];
+          const childrenAndYouth = [];
+          const familiesAndCommunities = [];
           const youth = [];
-          const familiesCommunities = [];
-          const nA = [];
+          const elderly = [];
+          const personsWithDisability = [];
+          const women = [];
+          const specialGroup = [];
+          const womenAndChildren = [];
+          const indigenousPeoples = [];
+          const drugDependents = [];
 
           response.data.forEach((item) => {
-            const clienteleName = item.Clientele;
+            const sector = item.Sector;
 
-            switch (clienteleName) {
+            switch (sector) {
               case "Children":
                 children.push(item);
                 break;
               case "Children and Youth":
-                youth.push(item);
+                childrenAndYouth.push(item);
                 break;
               case "Families and Communities":
-                familiesCommunities.push(item);
+                familiesAndCommunities.push(item);
                 break;
-              case "N/A":
-                nA.push(item);
+              case "Youth":
+                youth.push(item);
+                break;
+              case "Elderly":
+                elderly.push(item);
+                break;
+              case "Persons with Disability":
+                personsWithDisability.push(item);
+                break;
+              case "Women":
+                women.push(item);
+                break;
+              case "Special Group":
+                specialGroup.push(item);
+                break;
+              case "Women and Children":
+                womenAndChildren.push(item);
+                break;
+              case "Indigenous Peoples":
+                indigenousPeoples.push(item);
+                break;
+              case "Drug Dependents":
+                drugDependents.push(item);
                 break;
               default:
                 // Handle other cases if necessary
@@ -759,14 +793,249 @@ export default {
 
           // Calculate data lengths
           const childrenLength = children.length;
+          const childrenAndYouthLength = childrenAndYouth.length;
+          const familiesAndCommunitiesLength = familiesAndCommunities.length;
           const youthLength = youth.length;
-          const familiesCommunitiesLength = familiesCommunities.length;
-          const nALength = nA.length;
+          const elderlyLength = elderly.length;
+          const personsWithDisabilityLength = personsWithDisability.length;
+          const womenLength = women.length;
+          const specialGroupLength = specialGroup.length;
+          const womenAndChildrenLength = womenAndChildren.length;
+          const indigenousPeoplesLength = indigenousPeoples.length;
+          const drugDependentsLength = drugDependents.length;
 
-          console.log(childrenLength);
-          console.log(youthLength);
-          console.log(familiesCommunitiesLength);
-          console.log(nALength);
+          // Prepare and return data
+          const sectordata = {
+            labels: [
+              "Children",
+              "Children and Youth",
+              "Families and Communities",
+              "Youth",
+              "Elderly",
+              "Persons with Disability",
+              "Women",
+              "Special Group",
+              "Women and Children",
+              "Indigenous Peoples",
+              "Drug Dependents",
+            ],
+            label: ["Regional Operation"],
+            values: [
+              childrenLength,
+              childrenAndYouthLength,
+              familiesAndCommunitiesLength,
+              youthLength,
+              elderlyLength,
+              personsWithDisabilityLength,
+              womenLength,
+              specialGroupLength,
+              womenAndChildrenLength,
+              indigenousPeoplesLength,
+              drugDependentsLength,
+            ],
+            backgroundColor: [
+              "rgba(19, 63, 92, 1)",
+              "rgba(255, 0, 0, 1)", // Red
+              "rgba(0, 255, 0, 1)", // Green
+              "rgba(0, 0, 255, 1)", // Blue
+              "rgba(255, 255, 0, 1)", // Yellow
+              "rgba(255, 0, 255, 1)", // Magenta
+              "rgba(0, 255, 255, 1)", // Cyan
+              "rgba(128, 128, 128, 1)", // Gray
+              "rgba(255, 165, 0, 1)", // Orange
+              "rgba(0, 128, 0, 1)", // Dark Green],
+            ],
+          };
+          // Set barChartData to the computed data
+          this.SectorData = sectordata;
+        })
+        .catch((error) => {
+          console.error("Error fetching data:", error);
+
+          // Prepare and return data
+          const sectordata = {
+            labels: [
+              "Children",
+              "Children and Youth",
+              "Families and Communities",
+              "Youth",
+              "Elderly",
+              "Persons with Disability",
+              "Women",
+              "Special Group",
+              "Women and Children",
+              "Indigenous Peoples",
+              "Drug Dependents",
+            ],
+            label: ["Regional Operation"],
+            values: [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
+            backgroundColor: [
+              "rgba(19, 63, 92, 1)",
+              "rgba(255, 0, 0, 1)", // Red
+              "rgba(0, 255, 0, 1)", // Green
+              "rgba(0, 0, 255, 1)", // Blue
+              "rgba(255, 255, 0, 1)", // Yellow
+              "rgba(255, 0, 255, 1)", // Magenta
+              "rgba(0, 255, 255, 1)", // Cyan
+              "rgba(128, 128, 128, 1)", // Gray
+              "rgba(255, 165, 0, 1)", // Orange
+              "rgba(0, 128, 0, 1)", // Dark Green],
+            ],
+          };
+          // Set barChartData to the computed data
+          this.SectorData = sectordata;
+        });
+    },
+
+    // ! NEED FIXING BEFORE DEPLOYING
+    ClienteleFetchData() {
+      return axios
+        .get("http://127.0.0.1:8000/api/clientele")
+        .then((response) => {
+          // Initialize arrays for active and expired agencies
+          const childrenclientele = [];
+          const familiesAndCommunitiesclientele = [];
+          const childrenAndYouthclientele = [];
+          const childrenYouthFamiliesCommunities = [];
+          const SeniorCitizensclientele = [];
+          const personsWithDisabilityclientele = [];
+          const indigenousPeoplesclientele = [];
+          const womenclientele = [];
+          const others = [];
+
+          response.data.forEach((item) => {
+            const client = item.Clientele;
+
+            switch (client) {
+              case "Children":
+                childrenclientele.push(item);
+                break;
+              case "Children (0-6 years old Abandoned, orphaned, surrendered or temporary sheltered)":
+                childrenclientele.push(item);
+                break;
+              case "Children and Youth":
+              case "Children (3-7 years old abandoned, neglected, orphaned) Youth":
+              case "Children (Male abandoned and neglected)":
+              case "Children (0-12 years old abandoned, neglected, foundling)":
+              case "Children (Children In Need of Special Protection/ CNSP)":
+              case "Children (Abandoned, neglected, dependent, orphaned)":
+              case "Children (Critical and Chronically-ill patients both cancer and non-cancer)":
+              case "Children (Abandoned, Indigent, In-difficult situation)":
+              case "Children (0-5 years old dependent, neglected, foundling, abandoned, orphaned)":
+              case "Children (Street, Girls), Parents, Families, Communities":
+              case "Children diagnosed with cancer (0-18 y/old)":
+              case "Children (0-8 years old abandoned, orphaned, neglected)":
+              case "Children and Youth (Poor, Indigent, Neglected)":
+                childrenAndYouthclientele.push(item);
+                break;
+              case "Children (Disadvantaged) Youth (Out-of-School/OSY) Families and Communities":
+              case "Children (With Special Needs)":
+              case "Children (Street, Neglected)":
+              case "Children (4-13 years old) Families and Communities":
+              case "Children and Youth (Children In Need of Special Protection/CNSP)":
+              case "Children and Youth (Disadvantaged) (Out-of-School-Youth/OSY) Communities":
+                childrenYouthFamiliesCommunities.push(item);
+                break;
+              case "Families and Communities":
+              case "Families (Disadvantaged/displaced)":
+              case "Poor, vulnerable, marginalized, disadvantaged Children, Individuals, Families, and Communities":
+              case "Disadvantaged and marginalized Children, Youth, Indigenous Peoples, Families, and Communities":
+              case "Marginalized, Disadvantaged Individuals, Families and communities, Children, Youth Women, Victims-survivors of natural and human-induced calamities/disasters":
+                familiesAndCommunitiesclientele.push(item);
+                break;
+              case "Senior Citizens":
+                SeniorCitizensclientele.push(item);
+                break;
+              case "Persons with Disability":
+                personsWithDisabilityclientele.push(item);
+                break;
+              case "Indigenous Peoples":
+                indigenousPeoplesclientele.push(item);
+                break;
+              case "Women":
+              case "Women and Children":
+                womenclientele.push(item);
+                break;
+              case "Day Care Workers":
+              case "Drug Dependents":
+              case "Filipino seafarers, disadvantaged children, youth, families and communities":
+              case "N/A":
+                others.push(item);
+                break;
+              default:
+                others.push(item);
+                break;
+            }
+          });
+
+          // Calculate data lengths
+          const childrenclienteleLength = childrenclientele.length;
+          const familiesAndCommunitiesclienteleLength =
+            familiesAndCommunitiesclientele.length;
+          const childrenAndYouthclienteleLength =
+            childrenAndYouthclientele.length;
+          const childrenYouthFamiliesCommunitiesLength =
+            childrenYouthFamiliesCommunities.length;
+          const SeniorCitizensclienteleLength = SeniorCitizensclientele.length;
+          const personsWithDisabilityclienteleLength =
+            personsWithDisabilityclientele.length;
+          const indigenousPeoplesclienteleGroupLength =
+            indigenousPeoplesclientele.length;
+          const womenclienteleLength = womenclientele.length;
+          const othersLength = others.length;
+
+          console.log("TEST START");
+          console.log(childrenclienteleLength);
+          console.log(familiesAndCommunitiesclienteleLength);
+          console.log(childrenAndYouthclienteleLength);
+          console.log(childrenYouthFamiliesCommunitiesLength);
+          console.log(SeniorCitizensclienteleLength);
+          console.log(personsWithDisabilityclienteleLength);
+          console.log(indigenousPeoplesclienteleGroupLength);
+          console.log(womenclienteleLength);
+          console.log(othersLength);
+          console.log("TEST END");
+
+          // Prepare and return data
+          const clientdata = {
+            labels: [
+              "Children",
+              "Families and Communities",
+              "Children and Youth",
+              "Children Youth and Families",
+              "Senior Citizens",
+              "Persons With Disability",
+              "Indigenous Peples",
+              "Women",
+              "Others",
+            ],
+            label: ["Clientele"],
+            values: [
+              childrenclienteleLength,
+              familiesAndCommunitiesclienteleLength,
+              childrenAndYouthclienteleLength,
+              childrenYouthFamiliesCommunitiesLength,
+              SeniorCitizensclienteleLength,
+              personsWithDisabilityclienteleLength,
+              indigenousPeoplesclienteleGroupLength,
+              womenclienteleLength,
+              othersLength,
+            ],
+            backgroundColor: [
+              "rgba(19, 63, 92, 1)",
+              "rgba(255, 0, 0, 1)", // Red
+              "rgba(0, 255, 0, 1)", // Green
+              "rgba(0, 0, 255, 1)", // Blue
+              "rgba(255, 255, 0, 1)", // Yellow
+              "rgba(255, 0, 255, 1)", // Magenta
+              "rgba(0, 255, 255, 1)", // Cyan
+              "rgba(128, 128, 128, 1)", // Gray
+              "rgba(255, 165, 0, 1)", // Orange
+              "rgba(0, 128, 0, 1)", // Dark Green],
+            ],
+          };
+          // Set barChartData to the computed data
+          this.ClientData = clientdata;
         })
         .catch((error) => {
           console.error("Error fetching data:", error);
@@ -778,6 +1047,9 @@ export default {
     this.ClusterFetchData();
     this.RegionFetchData();
     this.AgencyFetchData();
+    this.SectorFetchData();
+
+    // ! NEED FIXINGS BEFORE DEPLOYING
     this.ClienteleFetchData();
   },
 };